+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Select `display_field_select` fields when requesting related data
 * Make local-filtering in `al-select` elements case insensitive
 * Fix `al-select` never getting more than the first page in `getRemoteFetchConfig` method
+* Add `hasOneParent` field support
 
 ## 0.2.2 (2023-01-23)